Create DashBoard page with the PowerBI example (having bug))
Create DashBoard page to embed the PowerBI example into React application. Following Microsoft documentation and README.md and yet still having major bug.

As the bug is not fixing yet, this page is all comment out as it wont affect the normal developing.

In the following days I will keep trying to debug.

The bug and problem will be sent into Slack channel, if anyone is find out the solution pls have a try and DM me.